+++ backend/internal/models/due_date.go
@@ -4,5 +4,5 @@ import "time"
 
 type DueDate struct {
 	Due          time.Time `json:"due"`
-	AssignmentID int64     `json:"assignment_id"`
+	AssignmentID int32     `json:"assignment_id"`
 }
